Add files via upload
This  project aims to develop an efficient and reliable keylogger solution for enhancing cybersecurity measures. The keylogger software will be designed to discreetly monitor and record keystrokes on a target device, providing valuable insights into user activity and potential security threats. With its robust functionality and user-friendly interface, the keylogger project offers an essential tool for researchers, cybersecurity professionals, and individuals seeking to bolster their defenses against unauthorized access and data breaches. Through continuous improvement and adaptation to evolving cybersecurity landscapes, this project strives to contribute to the ongoing efforts in safeguarding digital assets and ensuring privacy and security in the digital age.